Don't fail with an exception when can't connect to MSSQL on startup
Since most of the time Hangfire schema is already installed/updated, we shouldn't prevent application from starting up when there are connectivity issues with MSSQL instance.
